domain/requirements.txt
grpcio==1.68.1
grpcio-tools==1.68.1
pika==1.3.2
pymongo==4.10.1